feat: add an owner arg to py_image_layer (#545)
I needed to be able to change the owner of some files in the final oci
image. `mtree_mutate` seemed the right way to do it, but I couldn't
control the mtree while using `py_image_layer` , so this adds an `owner`
option to handle the `mtree_mutate`

### Changes are visible to end-users: yes

- Searched for relevant documentation and updated as needed: yes
- Breaking change (forces users to change their own code or config): no
- Suggested release notes appear below: no

### Test plan

<!-- Delete any which do not apply -->

I tested by adding an `owner` attribute to an image I was generating.
I did not update any tests, but if this change is approved, I can go
back and add if we think its necessary.